refactor: move initScore initMainTheme to  utils
