+++ src/utils/index.tsx
@@ -1,3 +1,5 @@
 export { default } from './getRandomCoord';
 export { default as printMatrix } from './printMatrix';
 export { default as delay } from './delay';
+export { default as initMainTheme } from './initMainTheme';
+export { default as initScore } from './initScore';
